# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Category:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        items:
          type: integer
        type: array
      id:
        type: string
      name:
        type: string
      tenant_id:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Class:
    properties:
      capacity:
        type: integer
      category:
        $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Category'
      category_id:
        type: string
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        items:
          type: integer
        type: array
      id:
        type: string
      name:
        type: string
      price:
        description: 'Financial Data: Always use int64 for money, price, balance,
          or amounts'
        type: integer
      tenant_id:
        type: string
      type:
        description: '''private'' or ''group'''
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.CreateReportRequest:
    properties:
      enrollment_id:
        type: string
      evaluation_notes:
        type: string
      score:
        maximum: 100
        minimum: 0
        type: number
      title:
        type: string
    required:
    - enrollment_id
    - title
    type: object
...
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Enrollment:
    properties:
      class:
        $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Class'
      class_id:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: string
      joined_at:
        type: string
      status:
        description: '''pending'', ''active'', ''completed'', ''dropped'''
        type: string
      student:
        $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Student'
      student_id:
        type: string
      tenant_id:
        description: Cross-service, ordinary UUID
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.HTTPResponse:
    properties:
      data: {}
      message:
        example: Operation completed successfully
        type: string
      status:
        example: success
        type: string
    type: object
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Pagination:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
...
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Report:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      enrollment:
        $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Enrollment'
      enrollment_id:
        description: Relation to Enrollment
        type: string
      evaluation_notes:
        type: string
      id:
        type: string
      reporter_id:
        description: Cross-service, ordinary UUID
        type: string
      score:
        description: Decimal using float64
        type: number
      tenant_id:
        description: Cross-service, ordinary UUID
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ReportListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Report'
        type: array
      pagination:
        $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Pagination'
    type: object
...
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Student:
    properties:
      created_at:
        type: string
      date_of_birth:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      full_name:
        type: string
      id:
        type: string
      parent_id:
        description: Cross-service, ordinary UUID
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.UpdateReportRequest:
    properties:
      evaluation_notes:
        type: string
      score:
        maximum: 100
        minimum: 0
        type: number
      title:
        type: string
    required:
    - title
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
info:
  contact: {}
  description: Academic Management Service for KelolaKelas Platform
  title: KelolaKelas Academic Service API
  version: "1.0"
paths:
...
  /api/v1/reports:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.HTTPResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ReportListResponse'
              type: object
      security:
      - BearerAuth: []
      summary: List reports
      tags:
      - Reports
    post:
      consumes:
      - application/json
      parameters:
      - description: Report payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.CreateReportRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.HTTPResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Report'
              type: object
      security:
      - BearerAuth: []
      summary: Create report
      tags:
      - Reports
  /api/v1/reports/{id}:
    delete:
      parameters:
      - description: Report UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.HTTPResponse'
      security:
      - BearerAuth: []
      summary: Delete report
      tags:
      - Reports
    get:
      parameters:
      - description: Report UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.HTTPResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Report'
              type: object
      security:
      - BearerAuth: []
      summary: Get report
      tags:
      - Reports
    patch:
      consumes:
      - application/json
      parameters:
      - description: Report UUID
        in: path
        name: id
        required: true
        type: string
      - description: Report payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.UpdateReportRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.HTTPResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Report'
              type: object
      security:
      - BearerAuth: []
      summary: Update report
      tags:
      - Reports
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"